+++ cherrypic-api/src/main/java/org/cherrypic/domain/auth/enums/OauthProvider.java
@@ -1,4 +1,4 @@
-package org.cherrypic.domain.auth.entity;
+package org.cherrypic.domain.auth.enums;
 
 public enum OauthProvider {
     KAKAO,
