+++ cherrypic-api/src/main/java/org/cherrypic/domain/auth/service/AuthService.java
@@ -2,7 +2,7 @@ package org.cherrypic.domain.auth.service;
 
 import org.cherrypic.domain.auth.dto.request.IdTokenRequest;
 import org.cherrypic.domain.auth.dto.response.SocialLoginResponse;
-import org.cherrypic.domain.auth.entity.OauthProvider;
+import org.cherrypic.domain.auth.enums.OauthProvider;
 
 public interface AuthService {
     SocialLoginResponse socialLoginMember(OauthProvider provider, IdTokenRequest request);
